requirements.txt
streamlit
pypdf
PyMuPDF
pandas
pywebview
pyinstaller
